pyproject.toml
[tool.poetry]
name = "humanloop"
version = "0.8.13"
description = ""
readme = "README.md"
authors = []
keywords = []

classifiers = [
    "Intended Audience :: Developers",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
    "Operating System :: POSIX",
    "Operating System :: MacOS",
    "Operating System :: POSIX :: Linux",
    "Operating System :: Microsoft :: Windows",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed"
]
packages = [
    { include = "humanloop", from = "src"}
]

[project.urls]
Repository = 'https://github.com/humanloop/humanloop-python'

[tool.poetry.dependencies]
python = "^3.9"
httpx = ">=0.21.2"
httpx-sse = "0.4.0"
pydantic = ">= 1.9.2"
pydantic-core = "^2.18.2"
typing_extensions = ">= 4.0.0"

opentelemetry-sdk = "^1.27.0"
opentelemetry-api = "^1.27.0"
opentelemetry-instrumentation-openai = "^0.33.3"
opentelemetry-instrumentation-cohere = "^0.33.3"
opentelemetry-instrumentation-anthropic = "^0.33.3"
opentelemetry-instrumentation-mistralai = "^0.33.3"

[tool.poetry.dev-dependencies]
mypy = "1.0.1"
pytest = "^7.4.0"
pytest-asyncio = "^0.23.5"
python-dateutil = "^2.9.0"
types-python-dateutil = "^2.9.0.20240316"
ruff = "^0.5.6"
python-dotenv = "^1.0.1"
openai = "^1.52.2"

[tool.pytest.ini_options]
testpaths = [ "tests" ]
asyncio_mode = "auto"

[tool.mypy]
plugins = ["pydantic.mypy"]

[tool.ruff]
line-length = 120


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
